Skip turn, sprites, and cut money
Added the skip turn ability for players to avoid zugzwang with the right gem. Added a new game over graphic, cut money and store for now. Was not having an easy time with it and not sure it adds anything to gameplay. Also cut Miser Edge, as it has no function without gold to hoard.